+++ src/main.c
@@ -143,6 +143,7 @@ main(int argc, char **argv)
 		extern void confmode_init(void);
 		extern void diffmode_init(void);
 		extern void gitmode_init(void);
+		extern void textmode_init(void);
 		extern void shmode_init(void);
 		extern void mdmode_init(void);
 		extern void mkmode_init(void);
@@ -160,6 +161,7 @@ main(int argc, char **argv)
 		confmode_init();
 		diffmode_init();
 		gitmode_init();
+		textmode_init();
 		shmode_init();
 		mdmode_init();
 		mkmode_init();